大塚不具合対応：（１）承認者が二人いる場合に、日別修正での承認状況の表示が不正　（２）承認状況が正しく取得できない

+++ nts.uk/uk.com/wk.ctx/nts.uk.ctx.workflow.infra/src/main/java/nts/uk/ctx/workflow/infra/repository/resultrecord/JpaAppRootConfirmRepository.java
@@ -396,7 +396,7 @@ public class JpaAppRootConfirmRepository extends JpaRepository implements AppRoo
 					}
 					
 					val framesOfPhase = frames.stream()
-							.filter(f -> f.get(1) == phaseOrder)
+							.filter(f -> ((BigDecimal)f.get(1)).intValue() == phaseOrder)
 							.collect(toList());
 					
 					if (framesOfPhase.isEmpty()) {
